refactor: migrating dev server to Vite
