+++ apps/namada-interface/.eslintrc.cjs
