+++ apps/namada-interface/package.json
@@ -48,7 +48,7 @@
     "watch": "webpack-cli --watch",
     "start": "webpack-dev-server",
     "start:proxy": "node ./scripts/startProxies.js",
-    "dev": "NODE_ENV=development yarn start",
+    "dev": "vite",
     "dev:local": "NODE_ENV=development NAMADA_INTERFACE_LOCAL=\"true\" yarn dev",
     "dev:proxy": "NAMADA_INTERFACE_PROXY=true && ./scripts/start-proxies.sh && yarn dev:local",
     "build": "NODE_ENV=production && yarn wasm:build && webpack-cli",
@@ -96,6 +96,7 @@
     "@types/react-qr-reader": "^2.1.4",
     "@types/redux-mock-store": "^1.0.3",
     "@types/styled-components": "^5.1.22",
+    "@vitejs/plugin-react": "^4.2.1",
     "autoprefixer": "^10.4.16",
     "babel-plugin-styled-components": "^2.0.3",
     "css-loader": "^6.7.3",
@@ -121,6 +122,8 @@
     "ts-node": "^10.9.1",
     "tsconfig-paths-webpack-plugin": "^4.1.0",
     "typescript-plugin-styled-components": "^2.0.0",
+    "vite": "^5.2.7",
+    "vite-tsconfig-paths": "^4.3.2",
     "webpack-bundle-analyzer": "^4.10.1",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.11.1"
