+++ apps/namada-interface/tailwind.config.cjs
