+++ packages/crypto/src/init-inline.ts
@@ -1,4 +1,6 @@
 import initWasm, { InitOutput } from "./crypto/crypto";
-import wasm from "./crypto/crypto_bg.wasm";
+
+// @ts-expect-error https://vitejs.dev/guide/features#fetching-the-module-in-node-js
+import wasm from "./crypto/crypto_bg.wasm?url";
 
 export const init: () => Promise<InitOutput> = async () => await initWasm(wasm);
