+++ packages/hooks/package.json
@@ -21,6 +21,7 @@
     "@namada/chains": "0.2.1",
     "@namada/types": "0.2.1",
     "@namada/utils": "0.2.1",
+    "isomorphic-dompurify": "^2.6.0",
     "react": "^17.0.2",
     "typescript": "^5.1.3"
   },
