+++ packages/hooks/src/useSanitizedParams.ts
@@ -1,18 +1,18 @@
-import { useParams, Params } from "react-router-dom";
-import { sanitize } from "dompurify";
+import { sanitize } from "isomorphic-dompurify";
+import { Params, useParams } from "react-router-dom";
 
 type ParamsT<ParamsOrKey> = Readonly<
   [ParamsOrKey] extends [string] ? Params<ParamsOrKey> : Partial<ParamsOrKey>
->
+>;
 
 export function useSanitizedParams<
-  ParamsOrKey extends string | Record<string, string | undefined> = string
-> (): ParamsT<ParamsOrKey> {
+  ParamsOrKey extends string | Record<string, string | undefined> = string,
+>(): ParamsT<ParamsOrKey> {
   const params = useParams<ParamsOrKey>();
   return Object.entries(params).reduce(
     (acc, [key, value]) => ({
       ...acc,
-      [key]: typeof value === 'string' ? sanitize(value) : value
+      [key]: typeof value === "string" ? sanitize(value) : value,
     }),
     {} as ParamsT<ParamsOrKey>
   );
