+++ packages/shared/src/init-inline.ts
@@ -1,4 +1,5 @@
 import initWasm, { InitOutput } from "./shared/shared";
-import wasm from "./shared/shared_bg.wasm";
+//@ts-expect-error https://vitejs.dev/guide/features#fetching-the-module-in-node-js
+import wasm from "./shared/shared_bg.wasm?url";
 
 export const init: () => Promise<InitOutput> = async () => await initWasm(wasm);
